MVP Architecture | Add InMemoryViewEventBus specification
